improve dfs-radix1sort.c
In the pass that follows a radix sort take advantage of the existing
knowledge of sections with identical starting character.

That removes one obvious inefficiency of the original implementation.

The other obvious inefficience (copying of y[] to wrk[] and then back) is
still here.

The speed of this version on my test case is within 1-2% from
dfs-radix2sort and dfs-radix4sort, but still seems consistently
on the slow side.