improve robustness of dfs-radix1sort.c
Modyfy algorithm to limit maximal depth of recursion to ceil(log2(N)).
It's more important for dfs-radix1sort.c than for the rest of my radix sort
variants because in this algorithm h[] array is a part of recursion stack.

This version is slightly slower with my test case then the privious version
but the difference is small.